requirements.txt
asgiref==3.3.1
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-jsonfield==1.4.1
django-rest-auth==0.9.5
django-token==0.1.3
djangorestframework==3.12.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.6.0
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
Quandl==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
td-ameritrade-python-api==0.3.4
tda-api==0.7.1
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
